Move leaderboard data into postprocessor for each empire
