+++ backend/engine/extraction/postprocessors/__init__.py
@@ -1,7 +1,9 @@
 from .comparison_processor import ComparisonProcessor
+from .leaderboard_processor import LeaderboardProcessor
 
 
 def build_postprocessor_list(isolation_layer):
     return [
-        ComparisonProcessor(isolation_layer)
+        ComparisonProcessor(isolation_layer),
+        LeaderboardProcessor(isolation_layer)
     ]
